fix(post): 홈화면 postcard 컴포넌트 로직 수정 및 채팅버튼 이동 - 홈화면의 postcard 컴포넌트를 클릭했을 때의 화면 전환을 로직 수정 - 채팅하기 버튼을 게시글 자세히보기 화면으로 이동
